# backend_typescript_service/src/public/swagger.json
{
	"openapi": "3.0.0",
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"ObjectId": {
				"type": "string",
				"description": "A class representation of the BSON ObjectId type."
			},
			"TodoStatus": {
				"enum": [
					"BACKLOG",
					"PENDING",
					"IN_PROGRESS",
					"COMPLETED"
				],
				"type": "string"
			},
			"Todo": {
				"properties": {
					"_id": {
						"$ref": "#/components/schemas/ObjectId"
					},
					"title": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"status": {
						"$ref": "#/components/schemas/TodoStatus"
					},
					"duration": {
						"type": "string"
					},
					"due_date": {
						"type": "string",
						"format": "date-time"
					},
					"created_at": {
						"type": "string",
						"format": "date-time",
						"default": "2026-01-29T06:39:54.341Z"
					},
					"updated_at": {
						"type": "string",
						"format": "date-time",
						"default": "2026-01-29T06:39:54.341Z"
					},
					"user": {
						"properties": {
							"$ref": {
								"type": "string"
							},
							"$id": {
								"$ref": "#/components/schemas/ObjectId"
							}
						},
						"required": [
							"$ref",
							"$id"
						],
						"type": "object"
					}
				},
				"required": [
					"_id",
					"title",
					"user"
				],
				"type": "object",
				"additionalProperties": false
			},
			"PaginatedTodoResponse": {
				"properties": {
					"items": {
						"items": {
							"$ref": "#/components/schemas/Todo"
						},
						"type": "array"
					},
					"total": {
						"type": "number",
						"format": "double"
					},
					"page": {
						"type": "number",
						"format": "double"
					},
					"size": {
						"type": "number",
						"format": "double"
					},
					"pages": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"items",
					"total",
					"page",
					"size",
					"pages"
				],
				"type": "object",
				"additionalProperties": false
			},
			"TodoCreateRequest": {
				"properties": {
					"title": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"status": {
						"$ref": "#/components/schemas/TodoStatus"
					},
					"due_date": {
						"type": "string"
					},
					"duration": {
						"type": "string"
					}
				},
				"required": [
					"title"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Record_string.number_": {
				"properties": {},
				"additionalProperties": {
					"type": "number",
					"format": "double"
				},
				"type": "object",
				"description": "Construct a type with a set of properties K of type T"
			},
			"Partial_TodoCreateRequest_": {
				"properties": {
					"title": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"status": {
						"$ref": "#/components/schemas/TodoStatus"
					},
					"due_date": {
						"type": "string"
					},
					"duration": {
						"type": "string"
					}
				},
				"type": "object",
				"description": "Make all properties in T optional"
			},
			"LoginResponse": {
				"properties": {
					"access_token": {
						"type": "string"
					},
					"token_type": {
						"type": "string"
					},
					"expires_in": {
						"type": "number",
						"format": "double"
					},
					"role": {
						"type": "string"
					},
					"name": {
						"type": "string"
					}
				},
				"required": [
					"access_token",
					"token_type",
					"expires_in",
					"role",
					"name"
				],
				"type": "object",
				"additionalProperties": false
			},
			"LoginRequestBody": {
				"properties": {
					"username": {
						"type": "string"
					},
					"password": {
						"type": "string"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"UserInfo": {
				"properties": {
					"_id": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"full_name": {
						"type": "string"
					},
					"role": {
						"type": "string"
					}
				},
				"required": [
					"_id",
					"email",
					"full_name",
					"role"
				],
				"type": "object",
				"additionalProperties": false
			},
			"VerifyRequestBody": {
				"properties": {
					"token": {
						"type": "string"
					}
				},
				"required": [
					"token"
				],
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {
			"jwt": {
				"type": "apiKey",
				"name": "Authorization",
				"in": "header"
			}
		}
	},
	"info": {
		"title": "express-ts-app",
		"version": "1.0.0",
		"description": "Express server with TypeScript for user authorization route",
		"license": {
			"name": "MIT"
		},
		"contact": {}
	},
	"paths": {
		"/todos": {
			"get": {
				"operationId": "GetTodos",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/PaginatedTodoResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Todos"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "page",
						"required": false,
						"schema": {
							"default": 1,
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "size",
						"required": false,
						"schema": {
							"default": 10,
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "sort_by",
						"required": false,
						"schema": {
							"default": "created_at",
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "sort_desc",
						"required": false,
						"schema": {
							"default": "true",
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "status",
						"required": false,
						"schema": {
							"$ref": "#/components/schemas/TodoStatus"
						}
					},
					{
						"in": "query",
						"name": "search",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "due_date_start",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "due_date_end",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "user_id",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"post": {
				"operationId": "CreateTodo",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Todo"
								}
							}
						}
					}
				},
				"tags": [
					"Todos"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/TodoCreateRequest"
							}
						}
					}
				}
			}
		},
		"/todos/stats/status": {
			"get": {
				"operationId": "GetStatsStatus",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Record_string.number_"
								}
							}
						}
					}
				},
				"tags": [
					"Todos"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": []
			}
		},
		"/todos/stats/workload": {
			"get": {
				"operationId": "GetStatsWorkload",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Todos"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": []
			}
		},
		"/todos/{id}": {
			"put": {
				"operationId": "UpdateTodo",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Todo"
								}
							}
						}
					}
				},
				"tags": [
					"Todos"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Partial_TodoCreateRequest_"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteTodo",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										}
									},
									"required": [
										"message"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Todos"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/health": {
			"get": {
				"operationId": "Health",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"status": {
											"type": "string"
										}
									},
									"required": [
										"status"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"General"
				],
				"security": [],
				"parameters": []
			}
		},
		"/user/profile": {
			"get": {
				"operationId": "GetProfile",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"tags": [
					"General"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": []
			}
		},
		"/admin/reports": {
			"get": {
				"operationId": "GetAdminReports",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"tags": [
					"General"
				],
				"security": [
					{
						"jwt": [
							"admin"
						]
					}
				],
				"parameters": []
			}
		},
		"/auth/login": {
			"post": {
				"operationId": "Login",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/LoginResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Auth"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/LoginRequestBody"
							}
						}
					}
				}
			}
		},
		"/auth/me": {
			"get": {
				"operationId": "GetMe",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UserInfo"
								}
							}
						}
					}
				},
				"tags": [
					"Auth"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": []
			}
		},
		"/auth/users": {
			"get": {
				"operationId": "GetUsers",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/UserInfo"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Auth"
				],
				"security": [
					{
						"jwt": [
							"admin"
						]
					}
				],
				"parameters": []
			}
		},
		"/auth/verify": {
			"post": {
				"operationId": "Verify",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"payload": {},
										"valid": {
											"type": "boolean"
										}
									},
									"required": [
										"payload",
										"valid"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Auth"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/VerifyRequestBody"
							}
						}
					}
				}
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}